% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grab.R
\name{.get_climate_shield_names}
\alias{.get_climate_shield_names}
\title{matching helper}
\usage{
.get_climate_shield_names(processing_units, type, climate_scenario)
}
\arguments{
\item{processing_units}{A \code{character} of the Processing Units, e.g. \code{'SpoKoot'}.}

\item{type}{A \code{character} indicating the type of fish, e.g. \code{'bt'}, \code{'ct'}.}

\item{climate_scenario}{A \code{character} indicating the climate scenario, e.g. \code{'baseline'}, \code{'moderate'}, \code{'extreme'}.}
}
\value{
A processing_units character.
}
\description{
matching helper
}
