% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grab.R
\name{.get_norwest_streamsegment_names}
\alias{.get_norwest_streamsegment_names}
\title{matching helper}
\usage{
.get_norwest_streamsegment_names(processing_units, type)
}
\arguments{
\item{processing_units}{A \code{character} of the Processing Units, e.g. \code{'SpoKoot'}.}

\item{type}{A \code{character} indicating the type of simple feature, e.g. \code{'points'}, \code{'lines'}.}
}
\value{
A processing_units character.
}
\description{
matching helper
}
