% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grab.R
\name{get_BullTroutNatalPatches}
\alias{get_BullTroutNatalPatches}
\title{Get Bull Trout Natal Habitat Patches}
\usage{
get_BullTroutNatalPatches()
}
\value{
A \code{sf} object.
}
\description{
This function calls Rocky Mountain Research Station (zip files) to get bull trout natal habitat patches
from \insertCite{isaak2022metapopulations}{fishguts} paper with associated model parameters.
}
\examples{
\dontrun{

# get R1 GRAIP-Lite
isaak2022 <- get_BullTroutNatalPatches()
}

}
