% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grab.R
\name{get_ClimateShield}
\alias{get_ClimateShield}
\title{Get Climate Shield scenarios}
\usage{
get_ClimateShield(
  processing_units,
  type = "bt",
  climate_scenario = "baseline",
  revised = FALSE,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{processing_units}{A \code{character} of the Processing Units, e.g. \code{'SpoKoot'}.}

\item{type}{A \code{character} indicating the type of fish, e.g. \code{'bt'}, \code{'ct'}.}

\item{climate_scenario}{A \code{character} indicating the climate scenario, e.g. \code{'baseline'}, \code{'moderate'}, \code{'extreme'}.}

\item{revised}{A \code{logical} for revised version. Likely to be deprecated as more information is known.}

\item{quiet}{A \code{logical}; suppress info on name, driver, size and spatial reference, or signaling no or multiple layers.}

\item{...}{Arguments to pass to `sf::read_sf()`.}
}
\value{
An \code{sf} object.
}
\description{
This function calls Rocky Mountain Research Station (zip files) to get Climate Shield
bull trout and cutthroat trout population occurrence scenarios for the western US \insertCite{isaak2017climate}{fishguts}.
}
\details{
The Climate Shield models developed by Isaak et al. (2015) provide stream-specific rangewide
 probabilistic predictions about the occurrence of juvenile Bull Trout and Cutthroat Trout under different
 scenarios of climate change and Brook Trout invasions. That information is available here as easy-to-use
 digital maps (.pdf files) and ArcGIS shapefiles for all streams within the historical ranges of these two
 native trout species across the western U.S. The geographic areas match the NorWeST production units because those stream temperature scenarios are integral
 to Climate Shield.

 The climate scenarios used in this project represent baseline (1980s), moderate (2040s), and extreme (2080s)
 climate change conditions and were chosen because they bracket what might be considered a “pristine” historical
 condition and “worst-case” end-of-century conditions.

\itemize{
\item  \strong{Clearwater River Basin} Includes data up to 2015
\item \strong{SpoKoot} Includes data up to 2015
\item \strong{Missouri Headwaters Unit} Includes data up to 2015
\item \strong{Upper Yellowstone-Bighorn} Includes data up to 2015
\item \strong{Upper Missouri-Marias}Includes data up to 2015
}
}
\examples{
\dontrun{

# get R1 GRAIP-Lite
spokoot_cs <- get_ClimateShield(processing_unit = 'SpoKoot', type = 'bt', climate_scenario = 'baseline')
}


}
\references{
{
\insertAllCited
}
}
\author{
Josh Erickson
}
