% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grab.R
\name{get_NorWestStreams}
\alias{get_NorWestStreams}
\title{Get NorWeST stream temperature scenarios}
\usage{
get_NorWestStreams(processing_units, type = "lines", quiet = TRUE, ...)
}
\arguments{
\item{processing_units}{A \code{character} of the Processing Units, e.g. \code{'SpoKoot'}.}

\item{type}{A \code{character} indicating the type of simple feature, e.g. \code{'points'}, \code{'lines'}.}

\item{quiet}{A \code{logical}; suppress info on name, driver, size and spatial reference, or signaling no or multiple layers.}

\item{...}{Arguments to pass to `terra::vect()`.}
}
\value{
An \code{sf} object.
}
\description{
This function calls Rocky Mountain Research Station (zip files) to get NorWest stream temperature data \insertCite{isaak2017norwest}{fishguts} for the month of August.
}
\details{
NorWeST stream temperature scenario maps were developed at a 1-kilometer resolution using spatial statistical stream network models.
 Stream temperature data used to fit the temperature model that created the scenario maps were screened using a consistent set of criteria to ensure the use of accurate temperature measurements.
Application of these techniques in previous research has yielded accurate and unbiased stream temperature models and predictions (R2 ~ 0.90; RMSE < 1.0 ˚C; for more details, click here).
 Multiple historical scenarios from 1993 to 2015 are available, as are future scenarios for mid-century (2030–2059), end-of-century (2070–2099), and date agnostic temperature increases
 (e.g., +1C relative to historical baseline periods). The stream temperature scenarios can be downloaded as geospatial data, which enables the display and querying of
 stream temperatures for river basins and other areas of interest across the western U.S.

\itemize{
\item  \strong{Eastern Montana} Includes data up to 2020
\item  \strong{Clearwater River Basin} Includes data up to 2015
\item \strong{SpoKoot} Includes data up to 2015
\item \strong{Missouri Headwaters Unit} Includes data up to 2015
\item \strong{Upper Yellowstone-Bighorn} Includes data up to 2015
\item \strong{Upper Missouri-Marias} Includes data up to 2015
\item \strong{Middle Columbia} Includes data up to 2015 and MWMT scenarios.
\item \strong{Salmon River Basin} Includes data up to 2015 and MWMT scenarios.
\item \strong{Snake-Bear} Includes data up to 2015 and MWMT scenarios.
\item \strong{Middle Snake} Includes data up to 2015 and MWMT scenarios.
}
}
\examples{
\dontrun{

# get SpoKoot NorWest polylines

NorWest_lines <- get_NorWestStreams(processing_units = 'SpoKoot', type = 'lines'))


}


}
\references{
{
\insertAllCited
}
}
\author{
Josh Erickson
}
